Rename htmlModalDemo to a generic reusable name
